+++ src/parameters/CraftliltpluginParameters.php
@@ -21,6 +21,7 @@ use lilthq\craftliltplugin\services\listeners\RegisterCpUrlRulesListener;
 
 class CraftliltpluginParameters
 {
+    public const REPORT_DATA = 'craft-lilt-plugin/get-report-data/invoke';
     public const TRANSLATION_REVIEW_ACTION = 'craft-lilt-plugin/translation/post-translation-review/invoke';
     public const TRANSLATION_PUBLISH_ACTION = 'craft-lilt-plugin/translation/post-translation-publish/invoke';
     public const TRANSLATION_REVIEW_PATH = 'craft-lilt-plugin/job/translation/review';